Added db-quotas.sh
Won't get to this before I'm not at Nexcess any more, but adding it
for completeness sake. This was supposed to find databases that had
incorrect ownership and then sum the disk space used. Then this was
to see if that space would be safe to add to the user's disk quota
without causing them to go over quota and break things. Guess I won't
be finishing it now, but it was worth a shot once ...

So long and thanks for all the fish.